Moved /ethereum under /pages/api to access env vars
