--- .env-sample
@@ -1,6 +1,0 @@
-# Wallet mnemonic will come from your MetaMask seed phrase
-WALLET_MNEMONIC=""
-# Infura endpoint is provided after you make a project under your account
-INFURA_ENDPOINT=""
-# This is the txn (as a hash) returned from running the `deploy.js` script
-SUPPLY_CHAIN_TRANSACTIONS_ADDRESS=""
