+++ package.json
@@ -5,7 +5,6 @@
   "main": "index.js",
   "dependencies": {
     "@truffle/hdwallet-provider": "^1.2.0",
-    "dotenv": "^8.2.0",
     "ganache-core": "^2.13.1",
     "solc": "^0.4.26",
     "truffle": "^5.1.50",
